separate scoll fade nav effect to a sparate file
