+++ webpack.config.js
@@ -29,8 +29,6 @@ module.exports = {
       },
     ],
 
-
   },
-
 
 };
